[develop] -> Fully added Athenian transformations. Updated the structure of the 3D object

+++ src/struct/object3/free.c
@@ -2,7 +2,7 @@
 // Created by stf20 on 29.09.2022.
 //
 
-#include "struct.h"
+#include "object3.h"
 
 void free_object3(object3_t* object3) {
   if (!object3)
@@ -12,7 +12,7 @@ void free_object3(object3_t* object3) {
     free(object3->list_vertex3.vertexes);
 
   if (object3->list_polygon.size) {
-    for (int i = object3->list_polygon.size; i < object3->list_polygon.size; i++) {
+    for (unsigned long i = object3->list_polygon.size; i < object3->list_polygon.size; i++) {
       if (object3->list_polygon.polygons[i].size)
         free(object3->list_polygon.polygons[i].vertexes3_id);
     }
